service and test part3

+++ src/main/java/ru/skypro/homework/service/ImageService.java
@@ -25,7 +25,7 @@ public interface ImageService {
 
     void saveToDatabase(ImageDTO imageDTO, Path imagePath, MultipartFile file);
 
-    Optional<Image> findImageById(Long id);
+    Optional<Image> findImageById(Integer id);
 }
